+++ crates/affinidi-did-resolver/affinidi-did-resolver-methods/did-webvh/src/resolve/ssi_resolve.rs
@@ -3,14 +3,10 @@
 */
 
 use crate::{
-    DIDWebVHError, DIDWebVHState,
-    log_entry::{LogEntry, MetaData},
-    log_entry_state::{LogEntryState, LogEntryValidationStatus},
-    parameters::Parameters,
+    DIDWebVHState,
     url::{URLType, WebVHURL},
     witness::proofs::WitnessProofCollection,
 };
-use reqwest::Client;
 use ssi::{
     dids::{
         DIDMethod, DIDMethodResolver, Document,
@@ -25,7 +21,6 @@ use ssi::{
 use static_iref::iri_ref;
 use std::time::Duration;
 use tracing::{Instrument, Level, span, warn};
-use url::Url;
 
 pub struct DIDWebVH;
 
@@ -144,7 +139,6 @@ impl DIDMethodResolver for DIDWebVH {
                 .log_entries
                 .last()
                 .ok_or_else(|| Error::internal("No log entries found"))?;
-            let metadata = &last_entry_state.metadata;
 
             // Get the latest DID Document
             let document: Document =
@@ -178,84 +172,4 @@ impl DIDMethodResolver for DIDWebVH {
 
 impl DIDMethod for DIDWebVH {
     const DID_METHOD_NAME: &'static str = "webvh";
-}
-
-impl DIDWebVH {
-    // Handles the fetching of the file from a given URL
-    async fn download_file(client: Client, url: Url) -> Result<String, DIDWebVHError> {
-        client
-            .get(url.clone())
-            .send()
-            .await
-            .map_err(|e| DIDWebVHError::NetworkError(format!("url ({url}): {e}")))?
-            .text()
-            .await
-            .map_err(|e| {
-                DIDWebVHError::NetworkError(format!("url ({url}): Failed to read response: {e}"))
-            })
-    }
-
-    /// Handles all processing and fetching for LogEntry file
-    async fn get_log_entries(
-        url: WebVHURL,
-        client: Client,
-    ) -> Result<Vec<LogEntryState>, DIDWebVHError> {
-        let log_entries_url = match url.get_http_url(Some("did.jsonl")) {
-            Ok(url) => url,
-            Err(e) => {
-                warn!("Invalid URL for DID: {e}");
-                return Err(DIDWebVHError::InvalidMethodIdentifier(format!(
-                    "Couldn't generate a valid URL from the DID: {e}"
-                )));
-            }
-        };
-
-        let log_entries_text = Self::download_file(client, log_entries_url).await?;
-
-        let mut log_entries = Vec::new();
-        for line in log_entries_text.lines() {
-            let log_entry: LogEntry = serde_json::from_str(line).map_err(|e| {
-                DIDWebVHError::LogEntryError(format!(
-                    "Failed to parse log entry from line: {line}. Error: {e}"
-                ))
-            })?;
-
-            log_entries.push(LogEntryState {
-                log_entry: log_entry.clone(),
-                metadata: MetaData::default(),
-                version_number: log_entry.get_version_id_fields()?.0,
-                validation_status: LogEntryValidationStatus::NotValidated,
-                validated_parameters: Parameters::default(),
-            });
-        }
-
-        Ok(log_entries)
-    }
-
-    /// Handles all processing and fetching for witness proofs
-    async fn get_witness_proofs(
-        url: WebVHURL,
-        client: Client,
-    ) -> Result<WitnessProofCollection, DIDWebVHError> {
-        let witness_url = match url.get_http_url(Some("did-witness.json")) {
-            Ok(url) => url,
-            Err(e) => {
-                warn!("Invalid URL for DID: {e}");
-                return Err(DIDWebVHError::InvalidMethodIdentifier(format!(
-                    "Couldn't generate a valid URL from the DID: {e}"
-                )));
-            }
-        };
-
-        let proofs_raw = Self::download_file(client, witness_url).await?;
-
-        Ok(WitnessProofCollection {
-            proofs: serde_json::from_str(&proofs_raw).map_err(|e| {
-                DIDWebVHError::WitnessProofError(format!(
-                    "Couldn't deserialize Witness Proofs Data: {e}",
-                ))
-            })?,
-            ..Default::default()
-        })
-    }
 }
